Add adjacency matrix

+++ src/main/java/uab/cn/nearestneighbour/NearestNeighbour.java
@@ -11,20 +11,20 @@ public class NearestNeighbour {
         stack = new Stack<Integer>();
     }
 
-    public void tsp(int adjCitiesDistance[][]) {
+    public void tsp(double adjCitiesDistance[][]) {
         nNodes = adjCitiesDistance[1].length - 1;
         int[] visited = new int[nNodes + 1];
         visited[1] = 1;
         stack.push(1);
         int element, dst = 0 ,i;
-        int min = Integer.MAX_VALUE;
+        double min = Double.MAX_VALUE;
         boolean minFlag = false;
         System.out.print(1 + "\t");
 
         while (!stack.isEmpty()) {
             element = stack.peek();
             i = 1;
-            min = Integer.MAX_VALUE;
+            min = Double.MAX_VALUE;
             while (i <= nNodes) {
                 if (adjCitiesDistance[element][i] > 1 && visited[i] == 0) {
                     if (min > adjCitiesDistance[element][i]) {
